pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.22.0",
    "azure-storage-blob>=12.25.1",
    "deepgram-sdk==2.12.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "gensim>=4.3.3",
    "iso639-lang>=2.6.0",
    "langdetect>=1.0.9",
    "nltk>=3.9.1",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "pymssql>=2.3.4",
    "scikit-learn>=1.6.1",
    "sumy>=0.11.0",
    "textblob>=0.19.0",
]
